Use raw syscalls instead of relying on cgo+libc
Since Go 1.6, Go pointers to memory containing other Go pointers are not
allowed to be passed to `cgo` anymore. Unfortunately, this makes
straightforward implementations of the `iovec` pattern impossible.

But system calls seem to be exempt from this rule, and as go uses raw
system calls on FreeBSD anyway, the implementation follows this pattern
and uses raw system calls.

Signed-off-by: Florian Limberger <flo@purplekraken.com>