Provide new type for flags to GetParams and SetParams
The use of the `flags` parameter of the `GetParams` and `SetParams`
functions required the use of the `gojail/syscall` package. This package
is intended to be an implementation detail, that should not be exposed
to the regular `gojail` user. To encapsulate the constants from the
`syscall` package and provide a basic layer of type safety, add the
`Flags` type which provides the values from the `syscall` package.

Signed-off-by: Florian Limberger <flo@purplekraken.com>